pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vocseg"
version = "0.1.0"
description = "Graduate-level semantic segmentation experiment stack for Pascal VOC 2007."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "matplotlib>=3.8",
  "numpy>=1.26",
  "pandas>=2.2",
  "pillow>=10.0",
  "pyyaml>=6.0",
  "scipy>=1.11",
  "scikit-learn>=1.4",
  "torch>=2.2",
  "torchvision>=0.17",
]

[project.optional-dependencies]
models = [
  "timm>=0.9.16",
  "segmentation-models-pytorch>=0.3.3",
]
sam2 = [
  "sam2",
]

[dependency-groups]
dev = [
  "pytest>=8.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.uv]
package = true
